PyChain/requirements.txt
django 
pygments 
requests 
psycopg2 
pycryptodome 

